% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLAMRAfunctions.R
\name{INLAMRA.control}
\alias{INLAMRA.control}
\title{Control parameters for INLAMRA}
\usage{
INLAMRA.control(
  Mlon = 1,
  Mlat = 1,
  Mtime = 1,
  randomSeed = 24,
  nuggetSD = 1e-05,
  numKnotsRes0 = 20L,
  J = 2L,
  numValuesForIS = 100,
  numIterOptim = 25L,
  distMethod = "haversine",
  normalHyperprior = TRUE,
  numISpropDistUpdates = 0,
  tipKnotsThinningRate = 1,
  credIntervalPercs = c(0.025, 0.975),
  timeJitterMaxInDecimalDays = 1/8.64e+08,
  spaceJitterMax = 1e-06,
  numOpenMPthreads = 1L,
  saveData = TRUE,
  fileToSaveOptOutput = NULL,
  folderToSaveISpoints = NULL,
  IScompleted = FALSE
)
}
\arguments{
\item{Mlon}{Number of longitude splits used to create the nested resolutions.}

\item{Mlat}{Number of latitude splits used to create the nested resolutions.}

\item{Mtime}{Number of time splits used to create the nested resolutions.}

\item{randomSeed}{Seed for the random number generator used for jittering and knot placement. Takes value 24 by default.}

\item{nuggetSD}{A small number added to the diagonal of the covariance matrices obtained by applying the Matern formula, to ensure that they are invertible. Takes value 1e-5 by default.}

\item{numKnotsRes0}{Number of knots at resolution 0 (the resolution encompassing the entire spatiotemporal domain). We would not recommend setting it under 8, as knots are first placed on the vertices of a rectangular prism nested within each subregion. Increasing this number also increases the program's memory footprint and running time.}

\item{J}{Multiplier used to determine the number of knots at each resolution. The number of knots in each subregion at resolution i is ceiling(numKnotsRes0 * (J/2)^i). We do not recommend setting J under 2, as some regions might end up with only two knots when M is large enough.}

\item{numValuesForIS}{The number of IS samples to be used for marginalisation.}

\item{numIterOptim}{Number of iterations in the L-BFGS algorithm used to identify the maximum of the join marginal posterior distribution of the hyperparameters. Could be set somewhat lower, 20 say, to reduce running time, but setting it too low might create imbalance in the importance sampling weights.}

\item{distMethod}{String indicating the method used to obtain distances in kilometers from longitude/latitude coordinates. Takes value "haversine" by default, for the Haversine distance formula. No alternative is implemented for now.}

\item{normalHyperprior}{Should hyperparameters be modelled on the logarithmic scale and normal hyperpriors be used? Takes value TRUE by default. Modelling hyperparameters on the original scale, with gamma priors, is possible, but not recommended.}

\item{numISpropDistUpdates}{The number of importance sampling (IS) weight updates in the adaptive IS algorithm. Takes value 0 by default. We implemented an adaptive IS scheme in case a reasonable level of balance in IS weights was not reached. It is enabled by setting this pararameter to 1 or more.}

\item{tipKnotsThinningRate}{Proportion of observation spatiotemporal locations that should be used as knots at the finest resolution, values should be in (0, 1]. Takes value 1 by default. A lower value for this parameter could reduce predictive performance, but greatly reduce the memory footprint. For very large datasets, lower values of this parameter are recommended, and can even be necessary.}

\item{credIntervalPercs}{Quantile boundaries of the reported credibility intervals. By default, 0.025 and 0.975, for a 95\% credibility interval.}

\item{timeJitterMaxInDecimalDays}{The maximum jittering to apply to time values, in days. Default value is 1/864000000 (1e-5 seconds). This is used to make it possible to evenly split the data into subregions at any depth. A balanced distribution of observations in the different regions makes it easier to control the method's computational performance. Putting this value at 0 disables the time jittering, which is not recommended.}

\item{spaceJitterMax}{The maximum jittering to apply to longitude/ latitude coordinates. Takes value 1e-5 by default. The method might run into numerical difficulties if longitude/latitude coordinates are replicated. The jittering ensures that it does not happen. Putting this value at 0 disables the spatial jittering, which is not recommended.}

\item{saveData}{Logical. Indicates whether data used to fit the model and prediction dataset (if applicable) should be bundled with the output, which is helpful for plotting. Can be disabled if provided datasets are huge and memory is limited.}

\item{fileToSaveOptOutput}{Character. An optional file name indicating where the optimiser's output should be saved. Specifying it ensures that INLAMRA can resume after the optimisation step.}

\item{folderToSaveISpoints}{Character. An optional *folder* name indicating where the importance sampling weight values should be saved. Specifying it ensures that INLAMRA can properly resume at the IS iteration it had reached before being interrupted.}

\item{IScompleted}{Logical value indicating whether all importance sampling weights have been obtained. Takes value FALSE by default. Set it to TRUE to produce intermediate results (based on fewer IS iterations than had been originally planned) when the run takes too long time to finish. Note that control$fileToSaveOptOutput and control$folderToSaveISpoints must be specified for this feature to work.}
}
\value{
A list with all control parameters.
}
\description{
Some important and less important tuning parameters for the INLA-MRA algorithm. Use this function to modify control parameters.
}
\details{
Some of the control parameters should be tuned to ensure better computational or predictive performance: Mlon, Mlat, Mtime, numKnotsRes0, J, numValuesForIS, numIterOptim, tipKnotsThinningRate.
Other control parameters make it possible to stop an INLAMRA run at any point and resume close to where the algorithm initially stopped, or produce an intermediate results based on an incomplete run: fileToSaveOptOutput, folderToSaveISpoints, IScompleted.
There are other control parameters that users should not normally need to change: distMethod, nuggetSD, normalHyperprior, spaceJitterMax, timeJitterMaxInDecimalDays.
}
\examples{
\dontrun{
INLAMRA.control(Mlon = 2, Mlat = 2, Mtime = 2)
}
}
