% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.INLAMRA}
\alias{plot.INLAMRA}
\title{Plots the output from INLAMRA}
\usage{
\method{plot}{INLAMRA}(
  x,
  filename = NULL,
  type = c("joint", "training", "predictions", "SD", "marginals"),
  polygonsToOverlay = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{x}{INLAMRA object,}

\item{filename}{Character. File in which to save the output.}

\item{type}{Character. The default, "joint" creates a grid of plots, one plot per time value, showing training data incremented with the predictions. ""SD" creates a grid of plots showing standard deviations for predictions. "trainingData" and "predictions" produce a similar grid with only the training data and predictions, respectively. "pars" produces plots of the marginal posterior densities for the parameters and hyperparameters.}

\item{polygonsToOverlay}{Optional SpatialPolygons object. Those polygons, e.g. regional boundaries, are added to each prediction map.}

\item{control}{List of control options, cf. ?plot.control.}

\item{...}{Arguments for the graphics engine; 'width' and 'height' should be especially helpful.}
}
\value{
A list of the plotted raster/SpatialPoints objects if type is anything but "pars"; a list of matching x/y values for the plotted (hyper)parameter posteriors otherwise.
}
\description{
Plots the predictions or the marginal posteriors produced by INLAMRA. The predictions can only be plotted if the data are found in the INLAMRA object (it will be the case unless the user explicitly disabled it by setting control parameter saveData to FALSE).
}
\details{
The function produces a grid of rasters by default, one per time value. If your data are not spaced regularly, you can plot the result as points instead by setting `control = plot.control(plotRaster = FALSE)`. If you have data collected over more than a few days, selecting the days to plot with the `timeValues` option in control is recommended.
}
\examples{
\dontrun{
INPUT_AN_EXAMPLE()
}
}
